Add root myocamlbuild.
The ocamljs stuff is commented out because -pack doesn't work with it, but I've tested without building packs and I was able to build lib/std/js without issues.

Also Xen backend is broken, as we uses only os/unix/oS, but should be easy to change that to os/xen/oS (just need to update the reference before calling ocamlbuid). Mir can set a env variable that ocamlbuild will read for example.