[unix] rationalise the device model to support multiple types, and update the block regression test to reflect this.
Instead of hotplug functions being called, there is now a 'find' function that will block until the device becomes available, and a listen function to tell the user when a fresh device is plugged into the tree. With these two functions, a specific device can be found, or the user can wait for a generic block device to be plugged in.