Add support for device dependencies
The device manager now accepts a tuple of (id, id list) on the plug mvar, where the list is the list of dependent ids for the device. The create function to the provider is passed a list of entries of the (plugged in) devices, which it can use to form a dependency chain.

Note that unplug support is not yet included, but will need to take into account the dependency list when it is...