[build] also copy .mli files from base libraries to std/
The build rules currently copy in 'source packs' (via ocaml-ocp) from lib/*/foo.ml into lib/std/foo.ml, and they are built in std/. We were only copying the .ml file previously and not the .mli, and so it was only being inferred and not correctly restricted. The new build rule also copies in the .mli file, ensuring the correct interfaces are used as intended.

This change will require .mli files for most source packs now (hence the fixes in f0772acabcc for example)