[buil] support for mapping kv_ro devices and custom device linkage
The %.mir file has a slightly different format now. The first line is still the name of the main function of the program, but subsequent lines are additional modules that are forced to be linked to the final binary. This is needed to specify device providers that the program expects to use (e.g. Block.Simple) since those will not otherwise be referenced by the main function and so need to be manually specified.  This also potentially permits some cool custom device synthesis in the future at compile time.

This patch also adds support for kv_ro: flags which are passed through to mir-run to indicate read-only k/v devices.