[scripts] add support for the -kv_ro flag to mir-run
The -kv_ro flag is used to map in read-only k/v devices to the binary. The <cfg> argument is dependent on the backend being used. Currently only unix-socket is supported, where it is "id:root_dir", e.g. "foo:/home" will map in a k_v device id "foo" with all keys relative to "/home"