[xen] add device model interfaces
The Xen Devices module is currently a symlink to the UNIX one, as there is nothing OS-dependent there at all; all of the logic to hotplug is actually in the device providers (and so the Xen versions will use xenstore watches for netif/blkif to provide the functionality that is currently done via command line parsing in UNIX).

Also add a stub Env module to Xen to match the UNIX one, which should provide arguments to the VM via Xenstore.