[dhcp]: Respect subtle difference between siaddr and DHCP server_identifier option
RFC2131 says that the siaddr names the server which takes part in the next
step of the bootstrap process (e.g. downloading an OS image). In the case of
my dhcp server, this is 0.0.0.0 because there is no second stage.

RFC2131 further says that the DHCP server will identify itself via the DHCP
option -- so we should reflect the provided server_identifier back to the
server rather than using the siaddr.

With this change my mirage VM gets an IP address!