[xen] experimental fix for the page corruption issue
I think we added the finaliser to the 'page' and then proceeded to take
a reference to a 'subbitstring' which is a separate object from the GC's
point of view.

As a quick workaround, switch to performing a full copy of the relevant
data and then immediately free the Io_page.