[std] add a String.index_between stdlib
The String.index_from function scans until the very end of the string, but when using Bitstrings, we often need to limit a scan to the particular view of the Bitstring. This new index_between function accepts a specific length argument and only scans until that point.

This commit is required to fix issue #52