[xen] add Res.write_response (same code for both 32- and 64-bit)
The only padding difference between the 32- and 64-bit response
structures is the 64-bit one has 4 bytes on the end. This doesn't
matter because (i) the responses are smaller than the requests so
this extra padding doesn't cause the ring slot size to increase;
and (ii) there aren't any fields after the last bit of padding.