Add a specific MacOS X platform
This lets us add MacOS X-specific libraries for the platform where
it diverges from stock Unix.  The first such use is in the networking
layer, where MacOS X provides the `vmnet` framework.

This patch defaults the `mirage CLI` to use the `MacOSX` mode on
that platform, and adds a `--no-macosx` override to force use of
the Unix mode (for advanced users).

The end-user capability added by this patch is that userlevel bridged
networking is possible using `vmnet`, and does not require any
third-party kernel modules to be installed either.

Requires ocaml/opam-repository#3135 to have the new packages