Remove DEVICE.connect
When a module is functorised over a DEVICE it should only have the
ability to *use* devices it is given, not to connect to new ones.

For example, this functor takes an abstract FS implementation and
uses connect to make a second FS.t onto the same underlying block
device, probably leading to corruption:

module MyMod (F : FS) = struct
  let use f =
    let f2 = F.connect (F.id f)) in
    ...
end

This change also lets us get rid of the 'id' function from DEVICE.

The 'id' and 'error' types are still present to avoid breaking
existing code that uses 'with id = ...', but can be removed later.