Builds with separated Arpv4 module type & combinators.
In this design, ipv4 doesn't keep a record of the underlying ethif,
clock, or time at all; this is stuffed into the arpv4 record (and impl).
With this design, Arpv4 has to provide pass-through read and write so
the Ipv4 implementation can access the underlying Ethif for its own
reads and writes (since apparently there are network use cases besides
resolving IPs to MACs??).