xl: generate an .xl.in as well as an .xl file
The .xl.in contains configuration needed for the domain to boot, for
now this is the line 'builder = linux' (i.e. PV mode) as well as
@VARIABLES@ for resources (paths, memory size) which should be
customised by the sysadmin.

Deployments should base themselves on the .xl.in file, substituting the
necessary variables, such that future changes (e.g. HVM boot) are
transparent.

Signed-off-by: David Scott <dave@recoil.org>