Add connect_raw to avoid forcing inputs
For logging devices, we want to initialise as early as possible.
With this change, a logging device can use `connect_raw` instead of
`connect` to get the unforced inputs (typically, any devices it needs to
operate plus the unikernel to be monitored). It can then force these in
the right order.

Note: impl now accepts subclasses so that this patch does not require
any changes to mirage. The `connect` methods in mirage should ideally be
marked as `private` at some point.