Build with ocamlbuild -r
This is the default behaviour if you have a _tags or myocamlbuild.ml,
but not otherwise in case you're running from your home directory. In
our case, we know the current directory is a valid project and can
always have this on.

Avoids a long and misleading error that is otherwise printed after every
compilation failure telling the user to enable this.

Fixes #536.