Add `disconnect` to `FLOW`
`FLOW` already includes `close` which performs a graceful shutdown
of one half of a bidirectional connection (also known as a half-close
and written in Unix as `Unix.(shutdown fd Unix.SHUTDOWN_SEND)`).
For example in the Mirage tcpip stack, `close` will send a FIN
and wait for the FIN ACK and then continue accepting data in the
other direction. This is important for protocols like HTTP where a
common pattern is to `connect`, `write` a request and then `close`
(a.k.a. `Unix.(shutdown fd SHUTDOWN_SEND)`), and then `read` the
response.

This patch requires a `disconnect` function which performs an
ungraceful shutdown of both halves of a bidirectional connection.
`disconnect` waits until all local resources have been cleaned up
(within limits imposed by the platform). Users who are happy to
let resources be freed in the background are welcome to call
`disconnect` asychronously.

Notes regarding implementations:

- in the Mirage tcpip stack, connection state is freed in the
  background as connections close and are timed-out. This extra
  function allows us to (i) request the state be cleaned up sooner;
  and (ii) optionally wait for the cleanup to be complete.
- in a Unix shim, `disconnect` should call `Unix.close` (or its
  equivalent) which will drop any buffered data rather than try to
  resend it, perform a FIN/ACK handshake and return. Unfortunately
  the Unix socket API is limited and it's not possible for us to
  wait while the kernel manages the connection in the "TIME-WAIT"
  state. This means it's possible even for careful applications
  to run out of kernel resources.

Signed-off-by: David Scott <dave@recoil.org>