provide qubes target; allow network autoconfigure
Add the mode `qubes`, which generally behaves the same as `Xen` with a few
exceptions noted below.

For the `qubes` target, provide an always-on qubes gui
listener and qrexec daemon, which are
initialized at unikernel start when the unikernel is configured with -t
qubes.  In other configurations they are not invoked.

Provide and expose a default_qubesdb in mirage.mli .

make argv_qubes, which passes only valid runtime_keys to bootvar.
This will only pass those runtime arguments which are expected to
the Functoria runtime, and therefore to Cmdliner.  Passing unexpected
arguments to Functoria results in runtime crashes, so this is needed
to reliably boot in environments where the supplied arguments can't be
easily controlled, such as Qubes or an external cloud hosting service.

make and expose qubes_ipv4_stack; generic_stackv4 chooses it when target is qubes.
`qubes_ipv4_stack` uses the Qubesdb_ipv4 sublibrary of `mirage-qubes`
along with `default_qubesdb` to
figure out the correct ipv4 settings and build a stackv4 based on them.
`generic_stackv4` chooses `qubes_ipv4_stack` rather than socket-based,
statically assigned, or dhcp-configured ipv4 if `mirage configure` is
invoked with `-t qubes`.