FS: remove `format` and `Format_unknown`
`format` is similar to `connect` which we removed previously from the
common interface: it is likely to be implementation-specific (e.g.
number of inodes, number of superblocks etc etc)

Furthermore it seems strange to have to acquire a `FS.t` handle in order
to be able to format a filesystem. This requires that implementations
track whether the `FS.t` is formatted or not (e.g. with a Some or None)
and then all operations may return `Format_unknown` (which they currently
don't). Removing `format` and `Format_error` fixes this problem.

With this change, `Fs.error` is a proper subset of `Fs.write_error`.

Signed-off-by: David Scott <dave@recoil.org>