Add an "ldpostflags" variable for Solo5 targets
Accounts for future flexibity in final link runes for Solo5 targets by
adding an "ldpostflags" pkg-config variable, the contents of which (if
present) are appended to the end of the final link command.

Also fixes a potential bug in the `pkg_config` callout function where
the result should be an empty list if the output is empty.