don't add "-color always" to ocamlbuild invocation, since it breaks some scripts.
OCaml (since 4.03.0) attempts to colorise warnings and errors (in a "auto" mode
by default, which checks whether stdout is a tty).  ocamlbuild does not execute
ocaml in a tty, this is why by default there is no color.

An upstream fix to consider the environment variable OCAML_COLOR in case no
"-color" command line flag was provided is proposed, and may be included in
4.05.0.  In the meantime, revert to no colored output (as in Mirage <3.0.0).

Fixes #797