RFC: Use ocamlfind and ocamlobjinfo to find Xen stubs
When OCaml is linking a normal binary, it discovers the `stubs.a` files
needed by the `library.cmxa` files and includes them. However when linking
a Xen/Qubes unikernel we control the final link ourselves, so we need
to know the list of `stubs.a` files to include.

The current behaviour is to add an extra key=value to the META file, e.g.:

  xen_linkopts = "-lio_page_xen_stubs"

and to build 2 versions of the stubs: one for Unix and one for Xen (sometimes
with different contents since the MiniOS APIs are non-standard)

This patch proposes to discover the `stubs.a` files from the OCaml archives,
as the OCaml compiler normally would, so we don't have to edit the META file.

We use the ocamlfind query suggested by @diml to list the archives from a
given ocamlfind package:

  ocamlfind query -r -format "%+a" -predicates native <package>

and then parse the output of ocamlobjinfo, for example:
```
$ ocamlobjinfo /home/opam/.opam/4.04.1/lib/io-page/io_page_xen.cmxa
File /home/opam/.opam/4.04.1/lib/io-page/io_page_xen.cmxa
Extra C object files: -lio_page_xen_stubs
Extra C options:
```

To distinguish between Unix stubs and Xen stubs, we would need to have a
convention like that used by io-page, i.e. insist that if a library `foo`
has Mirage-compatible stubs, it must have a subpackage called `foo.xen` which
will be queried and used for linking. This part is unimplemented.

Signed-off-by: David Scott <dave@recoil.org>