Automatically require io-page-{xen,unix} on Xen and Unix
Previously the correct stubs were found because `io-page` was in the
dependency list and everything was in the same `META` file with global
`xen_linkopts`. [mirage/io-page#46] splits `io-page` into 3 packages
and hence 3 distinct `META` files. Therefore we must depend explicitly
on `io-page-xen` to discover the `xen_linkopts`. This patch also adds
a dependency on `io-page-unix` to the `Unix` target for symmetry, although
there is no `xen_linkopts`: the OCaml compiler figures that out for itself.

Signed-off-by: David Scott <dave@recoil.org>