Fix cross-stage persistence of configure options
Previously, only 'if' keys were read from the cache (e.g. all the keys
which would reduce the application configuration space). With this fix,
we still do one pass to read the 'if' keys from the cache and reduce
the configuration. We added one more step where we get all the useful
keys in that reduced configuration, and read their value in the cache.

fix #116

Thanks to @Drup for the useful comments.