Add upper bounds to all packages specified in mirage config keys
This just adds an upper bound to the next major revision of the
relevant library.  This will make us slightly more future proof
against opam package changes, but will require more frequent
releases of this repository to relax upper bounds.  The only other
major source of incompatibilities coming up will be package
renaming (e.g. mirage-conduit to conduit-mirage), but that is
easier as it is a new package and will not require constraints.

This PR also fixes the conduit-mirage package and mirage/mirage-skeleton#232