Renaming for Solo5 0.4.0 release
Solo5 0.4.0 will rename the "ukvm" target to "hvt", and the Solo5 OPAM
packages to "solo5-bindings-TARGET". Adapt the mirage front-end tool to
match the new names.

Note that some references to "ukvm" build artifacts are kept to allow a
"mirage clean" to succeed after reconfiguring a dirty "ukvm" unikernel
source tree with the "hvt" target.