Update mirage_impl_random to expect an argument/implementation, an abstraction of entry-points of the underlying target.
This specific patch wants to provide a more conventional way (a
_functor_ way) to generate random generator according an entropy
engine which should register something at the beginning of the
procress. This is why random implementations expect an
entry-points abstraction.

This patch removes a nocrypto implementation which is a dummy
implementation to check if, at the link time, `Entropy` module is
correctly available - and avoid a linking error.