Emit dune's configuration for Solo5 target
This patch loads several files:
- libs.sexp used by `dune` when we want to provide main.exe.o (with
  -output-complete-obj)

  It will add right libasmrun.a, libnolibc.a and libopenlibm.a with
  the right patch (with -I) and expects libmirage-solo5_bindings.a in
  sources to link with it (and resolve some symbols at this step)

- ldflags used by `dune` and `ld` when we want to provide %{target}.hvt.

  It will use the linking script of Solo5 with some options like
  -nostdlib and max-page-size

- cflags.sexp to be able to compile correctly manifest.c and give the
  right context to compile any C artifacts inside the dune workspace

All of these files come from `ocaml-freestanding` and
libmirage-solo5_bindings.a comes from `mirage-solo5`. A `mirage
configure` does need to have this libraries but a `mirage build` needs
them - `make depends` serves this point.

These files want to replace `pkg-config`. The call of this tool is not
really deleted when it's done at the build time of `ocaml-freestanding`.
However, the mirage tool does not depend on `pkg-config` to get flags
and trust on the layout of `ocaml-freestanding` available on OPAM -
`dune` has the capability to resolve that.

* A rule is done to compile manifest.o needed at the final link time. A
  hack is done (we create a new file manifest.ml) to put as a dependency
  of %{target}.hvt manifest.o.

* A rule is done to do the final link step when we use the linking script
  from Solo5 and make the unikernel.