[Mirage 3.x] Port to new Mirage/Xen platform stack
This adapts the `mirage` front-end tool to the new Mirage/Xen platform
stack (see mirage/mirage#1159).

User-visible changes:

- `.xe` (Xenserver) filegeneration has been removed, I have no way to
test this.
- "Unknown" argument filtering in `Bootvar` / `argv` has been removed,
this was causing problems and I do not know of any users apart from
Qubes OS where this can be solved differently, see comments in
mirage/mirage#1159.

Note that Xen unikernels now include a dummy Solo5 manifest to keep the
rest of the build system and Solo5 layers happy; this does not contain
any devices as they are not required and Xen device naming does not
follow Solo5 requirements.