redirect stderr to /dev/null when invoking git
As reported in #1188, in case no git remote is specified (or no branch
is configured), git will complain to stderr "fatal: No such remote 'origin'".
The code in question is used to include a src url into the generated opam
file on a best-effort basis, and already handles errors (e.g. the unikernel is
not developed in a git repository) by using a dummy url.