Makefile: check that a single opam file is present
In #1198 a failure was observed if there are spurious opam files present
in the current working directory. While we can't inside of mirage handle this
properly, we can protect the `opam install -y --deps-only .` (which installs
dependencies of all present opam files) by checking that exactly one opam
file is present in the current working directory.