Fix the pull step when the build directory is custom
The generated makefile rule was wrong, it was cd-ing into the build directory
thus making the location of the lock file incorrect (as it is relative to the
makefile position). Instead, we use the --root / -r option of opam-monorepo
to fetch the duniverse folder inside the build directory, without having to cd
into it.