Delete the production of an alias for Solo5 targets
Currently, the alias produced tells to `dune` to compile everything.
However, some artifact (like `js_of_ocaml` artifacts) can not be built
into the Solo5 context (it's a non-sense).

To prevent such situation and let the user to define some artifacts which
should not be buildable into the Solo5 context AND still be able to build
everything needed for the unikernel (including such artifacts), we must delete
the alias to let `dune` to produce Solo5 artifact (the unikernel) and required
artifacts such as block-devices available only into the default context.

An other solution was about a precise alias which explicitely notices `dune` to
build the unikernel and block-devices. However, we are not able to pre-compute
block-devices easily and get such information when we emit `dune` rules
for the project.