Be able to specify an other context for [etc] files
As explained into the documentation, it's a non-sense to try to use
cross-compilation context to generate [etc] files. Only [bin] really needs the
cross-compiler. We explicitely generate [etc] files with the host compiler and
the default context. By this way, we are able to produce some artifacts (such
as a block-device) with the host compiler even if we want to cross-compile.