Add extra_deps for docteur device
To be able to notice dune to pre-compile some files needed to craft a docteur
image, we must add a new argument for this device to notify which files must be
compiled before make the docteur image.

More technically, `dune` is not able to infer that a file which should
be produce into a `dir/` directory is required when we explicitly notice a
dependency on `(source_tree dir)`.

Another change, and to be able to produce our docteur image, we must
install it as an [etc] artifact instead of a [bin] artifact (which
should be computed into the cross-compilation context).