Introduce an ALPN client and Mimic
This PR rename the initial `git_client` used to create a witness which is able
to initiate a Git connection to `mimic` and be able to re-use the same instance
for an new introduced ALPN client (http/1.1 and h2).

The ALPN client is currently used by the Robur's opam-mirror unikernel and can
be used in several contexts when the user wants to do an HTTP (http/1.1 or h2)
requests with TLS.

The mimic device (specially the `mimic_happy_eyeballs` device) initiate an
happy_eyeballs instance which can be used and re-used by several devices such as
`git_client` devices or the new introduced `alpn_devices`.

This commit breaks some names and specially some unikernels like our Git example
into our mirage-skeleton repository. The main change is the renaming of
`git_happy_eyeballs` to `mimic_happy_eyeballs` when the second is not really
related to Git stuff (and used by something else such as our new introduced
`alpn_client`).