check in mirage configure for a version lower bound in config.ml
If the first line of config.ml contains:
(* mirage >= a.b.c *) -- an OCaml comment

Then, a.b.c.d is checked to be smaller or equal the current version. This allows
within configure to error out early with a reasonable error message if a too old
mirage is attempted to be used with a unikernel.

Addresses #1381